Day-17/requirements.txt
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0

